Refactor unit model
The UIComponents now only have the properties they can make use of. This 
means that the angular components use the CSS properties explicitly 
instead of having all kinds of useless CSS properties.
Also it is no longer needed to update the CSS styles via event because 
Angular's change detection takes care of that.