[editor] Massively refactor unit and page view
- ... mostly for dynamic positioning
- Refactor reordering of sections. They can now be moved by small
buttons to the side, which appear when hovering. The section edit mode
is no longer available. -> KISS
- Removed all toolbars. The props can now be set via the props panel,
which now has sections for pages, sections and elements. This is to have
more vertical space available for editing the unit.
- Sections can now (only) be added via the toolbox panel, where the
second tab is now exclusively for sections. Future templates are
supposed to be whole sections, which can also be added from there.
- Refactor CSS layouting (flex and co) to have better handling of
vertically growing page (via sections)
- Some preparation for an upcoming overhaul of selection stuff
(again :))