[editor] Rework drag and drop of dynamic sections
To make it work that the section itself can handle drop events, but
also have the canvas to handle drops when outside of the section, all
the allowed dropLists have to be connected. Because the lists are not
properly nested (see below), this needs to be done manually by IDs.
This list is given to the necessary dropLists to make it possible to
drop items not only into them but also any other connected dropLists.

Dynamic sections have droplists for the grid cells next to the actual
elements. Elements can not be children of the grid cells because they
can span over multiple cells.

Dynamic sections don't have a general drop area, like static sections.
They have grid placeholder elements which are droplists. Therefore they
have no parent dropList to add to the list but themselves.
Static elements only have the parent, which is added to the list.

Resizing in dynamic sections is handled by the section/element-overlays
themselves.

Ticket #38