[editor] Allow dragging new elements into position
Elements can be dragged from the toolbox and will be placed where they
are dropped.

This is done by HTML drag and drop instead of Material dragNdrop. This
is because the setup is a lot easier and uncomplicated. In the future
the other Material drag and drop may be refactored away completely.