[editor] Fix element double click not opening edit dialog
The element has been obstructed by the resizing droplist. This extra 
droplist is removed.
To keep the parent droplist from handling the resize drop though, the 
element itself is a droplist and cathced the drop event. It is doin 
nothing with it, because resizing is done while dragging not on drop.