[editor] Fix element deletion via Delete key
This is now done via listener directly on the element overlay. The 
overlay's div is made selectable with 'tabindex'. Then the key event is 
caught and delete logic started. The actual logic did not change and 
still deletes all selected elements not just the one catching the delete 
key event.