[editor] Refactor directory structure
The main components are right at top level and the directories below
mimic the data structure (unit->page->section).
The toolbar is only used in dev mode and does not belong directly to the
unit view.
The TextEditor is a complex component and deserves it's own directory
even though it is used within the unit-view. Might be refactored to its
own module later.