Refactor unit factory
No longer is async. For elements which need upfront source info (media), 
the media prompf is done in the unit service. Here we can keep it 
simple.
Also remove coordinates from factory as this is information that does 
not need to be known here. This is also handled in the unit service 
where applicable.