[editor] Allow settings values to null
This is necessary to reset elements like radio button group. We use null 
for that instead of undefined.

The check was in place before to make sure input fields only accept 
valid values (no text in a number field for example). Angular thinks a 
null value is valid so we have to do this extra check. This is now done 
directly in the template before calling the updateModel method.