Refactor elements to use the model's value
This change mainly helps the editor.
The idea is to make the element components have the correct value of
their model and don't rely on the formcontrol value. The formcontrol
can be hard to manipulate, especially in compound elements. Setting
the model value is way easier.

The formcontrols are kept and should not interfere.