[editor] Refactor cursors
- Remove grab cursor from the canvas as it is no longer draggable (was
  in an earlier version)
- Fix element visual overlaying the resize element. Now the actual
  element is hidden, while the resize is going on.

Element cursor are changed so that there is a 'grab' cursor on hover
and a 'grabbing' cursor while active (clicked but not dragged yet)
and while dragging.
The resize cursor always stays the same, as there is not alternative
active curser (like grab/grabbing) for this.

On the dynamic overlay the curser cannot be set while the element
is being dragged. I assume it is because of 'relative' positioning.
The element is taken out of the DOM and pointer events don't work,
including cursors. This is why the body needs to be cursor-styled.
Cursors are set dynamically in the component via CSS class.
When dragging starts it is set and removed when it ends.