[editor] Fix TextEditor bullet list extension
ProseMirror seems not to allow editing text nodes. Since a text node can 
be a child of an element being applied the bullet list to, it fails for 
that text element. It seems it was workingly correctly before. The error 
seen on the console was inconsequential since we don't want to apply 
attributes to pure text nodes anyway. This change at least removes this 
error.