Refactor Cloze element to allow clicking on element in the player
A new variable is introduced for CompoundComponents: 
"allowClickThrough". With this set components can set their children and 
overlays to allow clicking on child elements or not. In the editor this 
is not wanted as the elements internal structure is different and the 
resulting event target can not be reliably styled. There we need the 
(span) overlay.
In the player though elements should be clickable as usual. The player 
can completely ignore this setting as it is on true (allowing clicking) 
by default.