[editor] Add font size to lists in TextEditor
The font size is applied by setting the list style. This is a 
workaround, since the list extension does not support setting font size 
directly. It takes the selected font size (which is now saved ina 
variable) and applies it. This is indepent of the font size of the text 
in the list. The item font size is done individually with a span anyway.

Also font size and list style droplist now (re-)apply the value on click 
of an item. This allows updating font sizes and list styles without 
having to select another option first.

- Add a new variable to save the selected font size and reuse it for 
list styles.