Fix cloze element re-generating all child elements on load
The player is supposed to take all the elements already defined and 
generated in the editor. Therefore the call to generate all elements 
from the text is not a general thing but specialized for the editor.

Right now the editor does no longer generate the cloze parts on load but 
only when the text prop has been touched.