Fix drag and drop element cursor style (in player)
This is the same hack used in the editor for overlay elements. Here 
applied to the actual dnd elements seen in the player.

Explanation: The cursor can not be applied to the dragged elements as it 
is not positioned (at least that is my understanding). The workaround is 
to style the body (cursor) while anything is being dragged.