[editor] Fix properties panel with multiselect
The change of moving some properties to sub objects broke the creation 
of the combinedProperties object, which can't deal with nested objects. 
To circumvent that all elements are now flattened before being combined.
The connected property panels now again look for the properties on the 
element itself and not in sub-objects.