Remove text field and area appearance variants and fix remaining
The other  variants are basically the same as filled. The only
difference is the background color, which we are manilupating anyway.
So there is no reason to keep them.

Also for the background color directive different target elements are
needed for the different appearance variants. This is now possible.
As stylings of the old appearance are kept when changing the appearance
the styles need to be removed first.
In the moment the appearance value is changed the component is not
actually changed yet and there we need to wait for another Angular
change detection cycle, via timeout.