Move ID-Service to editor
Generally only the editor needs to generate new IDs. Before there was
need to have it in common because some elements (like compound elements)
needed to generte new Elements including IDs.
Now all ID stuff is handled by the UnitService and the elements classes
are handed the fully built elements. Unfortunately this means that some
logic has to move back from classes into the UnitService. To un-bloat
the UnitService a (and in the future some more) helper class is created
for this logic.

The most important thing is, that the UIElement no longer generates
an ID when created. The ID has to be created before and passed to the
constructor.