Fix build process for custom theme
The theme file is now project specific because Angular demands assets
to be part of the project.
It also differs in content: The editor theme is only applied to canvas
elements. For the player every checkbox and radio button is affected.

The components now again use the default color of 'accent'. The custom
theme now has the green as accent color and no longer primary.