[player] Fix disappearance of scrollable page content
When a page has longer scrollable dynamic content that consists of many
individual elements (Likert), the content was sometimes not displayed
after interacting with those elements. This happened in the maximized
browser display, especially in Firefox.
Probably due to the deep dynamic nesting the height to be displayed
could not be calculated correctly