[player] Fix unstable behavior when marking text
When the user exits selecting between paragraphs, the offset of the
end container is set to 0. In Firefox this always (in Chrome sometimes)
leads to a misinterpretation of the selection. Therefore, the offset
is manipulated.

When the user finishes selecting between paragraphs and the selection
happens from back to front, Firefox does not consider the start
container as a selected child node. Therefore, it is added to the
list of selected nodes at the beginning.