[editor] Make menus not close when clicking on them
Also focus the height input field upon opening the menu.
For the layout menu this does not work as checkboxes don't play nicely 
with menus.

In fact having anything but button in the menu is against the Material 
spec. But i still think this is the best way to achieve the usability we 
want.