Make cloze element children not directly clickable
Only the overlay can be clicked. This way the statre of the elements can 
not be changed which might be confused for actual value changes. Those 
can still only be done via properties panel.

Unfortunately we need another variable so the cloze elements knows when 
it runs in the editor or not.