Change all selector prefixes from "app" to "aspect"
This is done according to the Angular best practices and minimizes
namespace collisions when using other packages which might use the same
selectors.

Technically the selectors are not even becessary since we use an Angular
component factory to create element components but it is still useful
for the DOM element naming, for debugging purposes for example.