Refactor unit definition version and sanatization
The unit definition now has 2 meta-fields: 'type' and 'version'.
The former has the fixed value of 'aspect-unit-definition'.
The latters contains the version without any prefixes, i.e. '1.2.3'.

The sanatizer tries to read the versions of older formats as well and
uses the found version. This is important since the handlePlusOne
method must only run of a specific version (in the old format).

The sanatizer no longer shows any messages and therefore does not need
the MessageService anymore. The information about whether sanatization
happened or not is returned as second parameter for those that are
interested. The editor is, the player is not.

The editor API service now constructs the 'unitDefinitionType'
parameter from both values, resulting in something like
'aspect-unit-definition@1.2.3'.