Add software keyboard
This can be enabled for TextFields and TextAreas and shows a keyboard
at the bottom of the screen. This does NOT set the field to read-only
which is probably wanted in a situation where the system keyboard is not
to be used. For flexibility purposes it is not implicitly set and needs
to be set manually. The same goes for other input assistance settings,
which should not be used in addition to the keyboard but it is allowed.

The keyboard has an extra switch to additionally show french special
characters.