Overhaul CI and add continuous deployment
Build and deploy script are written in bash for now as the syntax is
far easier and cleaner. Might be changed in the future as this requires
a Linux environment with bash available.

Deployment happens as version 9.9.9, because Teststudio does not accept
players and editors without semantic versioning.

The Dockerfile is the image used for building and deploying the app.