Implement copy on drop for dropLists
With this setting elements are copied when being dropped to another 
list.

On drop it is also checked if the item-ID is already present in the 
list. If it is the drop event is silently discarded. This allows putting 
items back in the list without creating duplicate IDs.

Lists with this setting:
- do show a placeholder of the items being dragged. This way it is 
conveyed that the item will remain there after being dropped.
- don't show a placeholder when items are hovered over them, to avoid 
confusion with duplicate items.

Sorting within DropLists is disabled, since it causes moving items 
within the list. It has to be discussed if this feature is even needed.