[player] Fix  ExpressionChangedAfterItHasBeenCheckedError in TextArea
- Replace UpdateTextareaPipe with DynamicRowsDirective

The dynamic height of the textarea is calculated based on the available
width. However, accessing the width of the textarea often results in an
error.
For this reason, the calculation is now done using resize and change
events in conjunction with a timeout to wait for the textarea to render.