Unfinished: Attach real file-inputs to button elements for file upload
This is to improve accessibility by having the upload element in the DOM 
at all times (though hidden) instead of dynamically creating and adding 
it to the DOM via script.

Unclear if this worth the effort and is actually the correct way to do 
this. Might be better to use label elements instead of buttons.