Rework class constructors
In Player-context the constructors are to fail, when the given 
properties are not exactly as expected. Therefoore they are validated. 
When this validation fails, an environment variable is checked to 
determine if the element should be created anyway with defaults. This 
switch is off in player-context. But the Editor is not so strict and can 
create elements with incomplete parameters.

There is also a new VersionManager which checks given unit definitions 
for version-validity. Upon failing this check Player and Editor show 
error messages.
The Editor accepts a certain unit definition version, but shows a 
warning.

The Editor checks units for invalid references after loading. If any are 
found, they are removed and an info-panel is shown.