[editor] Rework interaction-enabled functionality
Making element components unclickable is now applied to all child 
elements. This is necessary because some Material-components have child 
elements with specific pointer-events settings, which makes it possbile 
to interact
with them even though the parent is not clickable.

It is also now set via class on a parent element instead of the 
component itself.

#524