[editor][text-editor] Make font size optional
There can be a conflict between the font size set in the editor and the
element styling property. The editor value takes precedence, but there
was no way to get back to the default behaviour of controlling the size
via property. It is now possible to set the value to the default "null"
again, which removes the inline font size styling, allowing the element
property to apply again.

#675