[editor] Prevent selection of compound children via section list
The element list on every section allows selection elements directly
from there. This does not work for compound children right now, because
it is rather hard to get the actual component for this. This may be
implemented in the future. For now just disable the possibility to avoid
causing errrors.

#916