[audio] Restore the incorrect positioning of audios
In many units, audio elements were positioned editorially. This was done based on an incorrect spacing of the audio elements. When this was corrected in version 2.11.0, the editorially specified positions of the audio elments proved to be incorrect. To avoid having to revise all units, the incorrect spacing of audio elements is being reintroduced.